resultpage/tensorflow/requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor
astunparse==1.6.3
backcall
blinker==1.4
brotlipy==0.7.0
cachetools
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle
cryptography
cycler==0.10.0
cytoolz==0.10.1
dask
decorator==4.4.2
Django==3.0.8
gast==0.3.3
google-auth
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio
h5py==2.10.0
idna
image==1.5.32
imagecodecs
imageio
importlib-metadata
ipython
ipython-genutils==0.2.0
jedi
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown
matplotlib
networkx==2.4
numpy
oauthlib==3.0.1
olefile==0.46
opt-einsum==0+untagged.61.gd905544.dirty
pandas
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow
prompt-toolkit
protobuf==3.12.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests
requests-oauthlib
rsa
scikit-image==0.17.2
scikit-learn
scipy
six
sqlparse==0.3.1
tensorboard
tensorboard-plugin-wit
tensorflow==2.2.0
tensorflow-addons==0.10.0
tensorflow-estimator
termcolor==1.1.0
threadpoolctl
tifffile
toolz==0.10.0
tornado==6.0.4
tqdm
traitlets==4.3.3
typeguard==2.9.1
urllib3==1.25.9
wcwidth
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
